Remove RequestData and add to akka-http-server instead
